analyzer: split semantic analysis into its own directory.
