+++ analyzer/statements/function_declaration.cpp
@@ -27,7 +27,6 @@
 #include "vapor/analyzer/statements/block.h"
 #include "vapor/analyzer/statements/function_declaration.h"
 #include "vapor/analyzer/variables/overload_set.h"
-#include "vapor/codegen/ir/function.h"
 #include "vapor/parser/lambda_expression.h"
 
 namespace reaver::vapor::analyzer
@@ -73,69 +72,6 @@ inline namespace _v1
     statement_ir function_declaration::_codegen_ir(ir_generation_context &) const
     {
         return {};
-    }
-
-    future<> function_declaration::_analyze(analysis_context & ctx)
-    {
-        _function = make_function("overloadable function",
-            nullptr,
-            {},
-            [=, name = _parse.name.string](ir_generation_context & ctx) {
-                return codegen::ir::function{ U"operator()",
-                    {},
-                    fmap(_argument_list,
-                        [&](auto && arg) { return get<std::shared_ptr<codegen::ir::variable>>(get<codegen::ir::value>(arg.variable->codegen_ir(ctx))); }),
-                    _body->codegen_return(ctx),
-                    _body->codegen_ir(ctx) };
-            },
-            _parse.range);
-        _function->set_name(U"operator()");
-        _overload_set->add_function(this);
-
-        auto initial_future = [&] {
-            if (_return_type)
-            {
-                return (*_return_type)->analyze(ctx).then([&] {
-                    auto var = (*_return_type)->get_variable();
-
-                    assert(var->get_type() == builtin_types().type.get());
-                    assert(var->is_constant());
-
-                    _function->set_return_type(dynamic_cast<type_variable *>(var)->get_value());
-                });
-            }
-
-            return make_ready_future();
-        }();
-
-        return initial_future.then([&] { return when_all(fmap(_argument_list, [&](auto && arg) { return arg.type_expression->analyze(ctx); })); })
-            .then([&] {
-                auto arg_variables = fmap(_argument_list, [&](auto && arg) -> variable * {
-                    arg.variable->set_type(arg.type_expression->get_variable());
-                    return arg.variable.get();
-                });
-
-                _function->set_arguments(std::move(arg_variables));
-
-                return _body->analyze(ctx);
-            })
-            .then([&] {
-                fmap(_return_type, [&](auto && ret_type) {
-                    auto explicit_type = ret_type->get_variable();
-                    auto type_var = static_cast<type_variable *>(explicit_type);
-
-                    assert(type_var->get_value() == _body->return_type());
-
-                    return unit{};
-                });
-
-                _function->set_body(_body.get());
-
-                if (!_function->return_type().try_get())
-                {
-                    _function->set_return_type(_body->return_type());
-                }
-            });
     }
 
     future<statement *> function_declaration::_simplify(simplification_context & ctx)
